requirements.txt
feedparser==6.0.11
anthropic>=0.49.0
jinja2==3.1.4
pyyaml==6.0.2
python-dotenv==1.0.1
requests==2.32.3
trafilatura>=1.12.0
beautifulsoup4>=4.12.0
lxml>=5.3.0
